requirements.txt
numpy==1.19.5
opencv-python==4.2.0.32
flask==1.1.2
requests==2.25.1
Pillow==8.1.2
scipy==1.5.4
jsonpickle==2.0.0